+++ Clp/src/ClpModel.cpp
@@ -3289,7 +3289,7 @@ ClpModel::getRowName(int iRow) const
           indexError(iRow, "getRowName");
      }
 #endif
-     int size = rowNames_.size();
+     int size = static_cast<int>(rowNames_.size());
      if (size > iRow) {
           return rowNames_[iRow];
      } else {
@@ -3309,7 +3309,7 @@ ClpModel::setRowName(int iRow, std::string &name)
      }
 #endif
      unsigned int maxLength = lengthNames_;
-     int size = rowNames_.size();
+     size_t size = static_cast<int>(rowNames_.size());
      if (size <= iRow)
           rowNames_.resize(iRow + 1);
      rowNames_[iRow] = name;
@@ -3326,7 +3326,7 @@ ClpModel::getColumnName(int iColumn) const
           indexError(iColumn, "getColumnName");
      }
 #endif
-     int size = columnNames_.size();
+     int size = static_cast<int>(columnNames_.size());
      if (size > iColumn) {
           return columnNames_[iColumn];
      } else {
@@ -3346,7 +3346,7 @@ ClpModel::setColumnName(int iColumn, std::string &name)
      }
 #endif
      unsigned int maxLength = lengthNames_;
-     int size = columnNames_.size();
+     int size = static_cast<int>(columnNames_.size());
      if (size <= iColumn)
           columnNames_.resize(iColumn + 1);
      columnNames_[iColumn] = name;
@@ -3359,7 +3359,7 @@ void
 ClpModel::copyRowNames(const std::vector<std::string> & rowNames, int first, int last)
 {
      unsigned int maxLength = lengthNames_;
-     int size = rowNames_.size();
+     int size = static_cast<int>(rowNames_.size());
      if (size != numberRows_)
           rowNames_.resize(numberRows_);
      int iRow;
@@ -3375,7 +3375,7 @@ void
 ClpModel::copyColumnNames(const std::vector<std::string> & columnNames, int first, int last)
 {
      unsigned int maxLength = lengthNames_;
-     int size = columnNames_.size();
+     int size = static_cast<int>(columnNames_.size());
      if (size != numberColumns_)
           columnNames_.resize(numberColumns_);
      int iColumn;
@@ -3391,7 +3391,7 @@ void
 ClpModel::copyRowNames(const char * const * rowNames, int first, int last)
 {
      unsigned int maxLength = lengthNames_;
-     int size = rowNames_.size();
+     int size = static_cast<int>(rowNames_.size());
      if (size != numberRows_)
           rowNames_.resize(numberRows_);
      int iRow;
@@ -3414,7 +3414,7 @@ void
 ClpModel::copyColumnNames(const char * const * columnNames, int first, int last)
 {
      unsigned int maxLength = lengthNames_;
-     int size = columnNames_.size();
+     int size = static_cast<int>(columnNames_.size());
      if (size != numberColumns_)
           columnNames_.resize(numberColumns_);
      int iColumn;
@@ -3649,7 +3649,7 @@ ClpModel::rowNamesAsChar() const
      char ** rowNames = NULL;
      if (lengthNames()) {
           rowNames = new char * [numberRows_+1];
-          int numberNames = rowNames_.size();
+          int numberNames = static_cast<int>(rowNames_.size());
           numberNames = CoinMin(numberRows_, numberNames);
           int iRow;
           for (iRow = 0; iRow < numberNames; iRow++) {
@@ -3689,7 +3689,7 @@ ClpModel::columnNamesAsChar() const
      char ** columnNames = NULL;
      if (lengthNames()) {
           columnNames = new char * [numberColumns_];
-          int numberNames = columnNames_.size();
+          int numberNames = static_cast<int>(columnNames_.size());
           numberNames = CoinMin(numberColumns_, numberNames);
           int iColumn;
           for (iColumn = 0; iColumn < numberNames; iColumn++) {
@@ -3886,8 +3886,8 @@ ClpModel::createCoinModel() const
      for (i = 0; i < numberRows_; i++) {
           char temp[30];
           strcpy(temp, rowName(i).c_str());
-          int length = strlen(temp);
-          for (int j = 0; j < length; j++) {
+          size_t length = strlen(temp);
+          for (size_t j = 0; j < length; j++) {
                if (temp[j] == '-')
                     temp[j] = '_';
           }
@@ -3896,8 +3896,8 @@ ClpModel::createCoinModel() const
      for (i = 0; i < numberColumns_; i++) {
           char temp[30];
           strcpy(temp, columnName(i).c_str());
-          int length = strlen(temp);
-          for (int j = 0; j < length; j++) {
+          size_t length = strlen(temp);
+          for (size_t j = 0; j < length; j++) {
                if (temp[j] == '-')
                     temp[j] = '_';
           }
